requirements.txt
streamlit
pytesseract
opencv-python
numpy
Pillow
requests
langchain
langchain_groq
duckduckgo-search
python-dotenv
